+++ src/test/kotlin/com/nowsprinting/intellij_mob/timer/TimerServiceTest.kt
@@ -2,7 +2,7 @@
  * Copyright 2020 Koji Hasegawa. Use of this source code is governed by the Apache 2.0 license that can be found in the LICENSE file.
  */
 
-package com.nowsprinting.intellij_mob.service
+package com.nowsprinting.intellij_mob.timer
 
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Assertions.assertFalse