+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ## Unreleased
-
+- Display timer on statusbar
 
 ## 1.0-alpha3
 - Implements mob start, next, done, reset